refactor: extract order model and socket integration
